+++ ShnergleApp/AnalyticsViewController.m
@@ -26,7 +26,7 @@
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
-    UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:[NSString stringWithFormat:@"Cell%d", indexPath.row]];
+    UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:[NSString stringWithFormat:@"Cell%ld", (long)indexPath.row]];
     if (indexPath.row == 0) {
         cell.imageView.image = [UIImage imageNamed:@"glyphicons_228_gbp"];
     } else if (indexPath.row == 1) {
